requirements.txt
tetpyclient
python-dotenv